Refactor quorum docker entrypoint creation to its own file
